pyproject.toml
[project]
name = "ares"
version = "1.0.0"
description = "Autonomous Diagnostic Engine for Maritime Fleet Reliability"
authors = [{ name = "Adi", email = "adi@maritime-ai.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.9.3",
    "pydantic~=2.11.9",
    "pydantic-settings~=2.10.1",
    "qdrant-client>=1.17.0",
    "pdfplumber~=0.11.4",
    "pdf2image==1.16.3",
    "pytesseract==0.3.10",
    "pillow==10.1.0",
    "typer==0.9.0",
    "pyyaml==6.0",
    "openai~=1.83.0",
    "langchain",
    "langchain-openai",
    "python-dotenv~=1.1.1"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.0",
    "ruff==0.1.9",
    "mypy==1.7.0"
]

[project.scripts]
ares = "ares.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
project_name = "ares"
